shift student module files
