+++ app-modules/student-data-model/src/Filament/Imports/StudentManageableImporter.php
@@ -34,7 +34,7 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\StudentRecordManager\Filament\Imports;
+namespace AdvisingApp\StudentDataModel\Filament\Imports;
 
 use AdvisingApp\StudentDataModel\Models\Student;
 use Illuminate\Support\Str;
